Rename model file to base_echem
